requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
djangorestframework==3.11.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
PyJWT==1.7.1
Pillow==7.2.0
whitenoise==5.1.0
static3==0.7.0
psycopg2==2.8.5